requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
aiohttp=3.8.3=pypi_0
aiohttp-cors=0.7.0=pypi_0
aiorwlock=1.3.0=pypi_0
aiosignal=1.2.0=pypi_0
alembic=1.8.1=pypi_0
antlr4-python3-runtime=4.8=pypi_0
anyio=3.6.2=pypi_0
asttokens=2.2.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.1.0=pypi_0
autopage=0.5.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
blessed=1.19.1=pypi_0
bottleneck=1.3.5=py38h7deecbd_0
brotlipy=0.7.0=py38h0a891b7_1004
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=5.2.0=pypi_0
category-encoders=2.2.2=pypi_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.14.6=py38ha65f79e_0
cfgv=3.3.1=pypi_0
cftime=1.5.1.1=py38hce1f21e_0
chardet=4.0.0=pypi_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.0.4=pypi_0
cliff=4.0.0=pypi_0
cmaes=0.9.0=pypi_0
cmd2=2.4.2=pypi_0
colorful=0.5.4=pypi_0
colorlog=6.7.0=pypi_0
comm=0.1.2=pypi_0
configs=3.0.3=pypi_0
contourpy=1.0.6=pypi_0
cryptography=37.0.2=py38h2b5fc30_0
cuda=11.6.2=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=11.8.86=0
cuda-libraries=11.6.2=0
cuda-libraries-dev=11.6.2=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=11.8.87=0
cuda-runtime=11.6.2=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.6.2=0
cuda-tools=11.6.2=0
cuda-visual-tools=11.6.2=0
cudatoolkit=11.6.0=hecad31d_10
curl=7.84.0=h5eee18b_0
cycler=0.11.0=pypi_0
datasets=2.7.1=pypi_0
debugpy=1.6.4=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.6=pypi_0
distlib=0.3.6=pypi_0
einops=0.3.0=pypi_0
entrypoints=0.4=pypi_0
executing=1.2.0=pypi_0
fastapi=0.86.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.8.0=pypi_0
flaml=1.0.14=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.10.4=h0708190_1
frozenlist=1.3.1=pypi_0
fsspec=2022.11.0=pypi_0
future=0.18.2=pypi_0
fuzzywuzzy=0.18.0=pypi_0
gdown=4.5.1=pypi_0
gds-tools=1.4.0.31=0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-api-core=2.10.2=pypi_0
google-auth=2.14.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
googleapis-common-protos=1.56.4=pypi_0
gpustat=1.0.0=pypi_0
greenlet=2.0.1=pypi_0
grpcio=1.43.0=pypi_0
h11=0.14.0=pypi_0
hdf4=4.2.13=h3ca952b_2
hdf5=1.10.6=h3ffc7dd_1
huggingface-hub=0.11.1=pypi_0
hydra-core=1.0.7=pypi_0
identify=2.5.12=pypi_0
idna=2.10=pypi_0
importlib-metadata=5.0.0=pypi_0
importlib-resources=5.10.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipdb=0.13.9=pypi_0
ipykernel=6.19.4=pypi_0
ipython=8.7.0=pypi_0
ipywidgets=8.0.3=pypi_0
jedi=0.18.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h166bdaf_1
jsonschema=4.17.0=pypi_0
jupyter-client=7.4.8=pypi_0
jupyter-core=5.1.0=pypi_0
jupyterlab-widgets=3.0.4=pypi_0
kiwisolver=1.4.4=pypi_0
krb5=1.19.2=hac12032_0
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
liac-arff=2.5.0=pypi_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcurl=7.84.0=h91b91d3_0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libnetcdf=4.8.1=h42ceab0_1
libnghttp2=1.46.0=hce63b2e_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.37=h21135ba_2
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=hf544144_3
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f98852_1
libzip=1.5.1=h8d318fa_1004
lightgbm=3.3.3=pypi_0
lightly=1.2.27=pypi_0
lightly-utils=0.0.2=pypi_0
littleutils=0.2.2=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
mako=1.2.3=pypi_0
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.0=pypi_0
matplotlib-inline=0.1.6=pypi_0
minio=7.1.12=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
msgpack=1.0.4=pypi_0
multidict=6.0.2=pypi_0
multiprocess=0.70.14=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.6=pypi_0
netcdf4=1.5.7=py38ha0f2276_1
nettle=3.6=he412f7d_0
nodeenv=1.7.0=pypi_0
nsight-compute=2022.3.0.22=0
numexpr=2.8.4=py38he184ba9_0
numpy=1.19.5=pypi_0
nvidia-ml-py=11.495.46=pypi_0
oauthlib=3.2.2=pypi_0
ogb=1.3.5=pypi_0
olefile=0.46=pyh9f0ad1d_1
omegaconf=2.0.6=pypi_0
opencensus=0.11.0=pypi_0
opencensus-context=0.1.3=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openml=0.13.0=pypi_0
openssl=1.1.1s=h7f8727e_0
optuna=2.8.0=pypi_0
outdated=0.2.2=pypi_0
packaging=21.3=pypi_0
pandas=1.1.5=pypi_0
parso=0.8.3=pypi_0
patsy=0.5.3=pypi_0
pbr=5.11.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=pypi_0
pip=22.2.2=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=2.5.2=pypi_0
prefetch-generator=1.0.1=pypi_0
prettytable=3.5.0=pypi_0
prometheus-client=0.13.1=pypi_0
prompt-toolkit=3.0.36=pypi_0
protobuf=3.19.6=pypi_0
psutil=5.9.3=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py-spy=0.3.14=pypi_0
pyarrow=6.0.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=1.10.2=pypi_0
pydeprecate=0.3.0=pypi_0
pygments=2.13.0=pypi_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=pypi_0
pyperclip=1.8.2=pypi_0
pyrsistent=0.19.2=pypi_0
pysocks=1.7.1=pyha2e5f31_6
pytdc=0.3.7=pypi_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.13.0=py3.8_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_0
pytorch-lightning=1.3.6=pypi_0
pytorch-mutex=1.0=cuda
pytorch-tabnet=3.0.0=pypi_0
pytorch-tabular=0.7.0=pypi_0
pytz=2022.6=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=24.0.1=pypi_0
ray=1.13.0=pypi_0
rdkit-pypi=2022.9.3=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.10.31=pypi_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rsa=4.9=pypi_0
sacremoses=0.0.53=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.9.3=pypi_0
seaborn=0.12.1=pypi_0
setuptools=63.4.1=py38h06a4308_0
setuptools-scm=7.1.0=pypi_0
six=1.16.0=pyh6c4a22f_0
smart-open=6.2.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlalchemy=1.4.43=pypi_0
sqlite=3.39.3=h5082296_0
stack-data=0.6.2=pypi_0
starlette=0.20.4=pypi_0
statsmodels=0.13.5=pypi_0
stevedore=4.1.0=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.11.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.5.1=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
torchaudio=0.13.0=py38_cu116
torchcontrib=0.0.2=pypi_0
torchmetrics=0.6.0=pypi_0
torchvision=0.11.2=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.8.0=pypi_0
transformers=4.21.1=pypi_0
typing_extensions=4.4.0=pyha770c72_0
urllib3=1.26.12=py38h06a4308_0
uvicorn=0.19.0=pypi_0
virtualenv=20.16.6=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.4=pypi_0
wilds=2.0.0=pypi_0
wildtime=1.1.3=pypi_0
xarray=2022.11.0=py38h06a4308_0
xgboost=1.7.0.post0=pypi_0
xmltodict=0.13.0=pypi_0
xxhash=3.1.0=pypi_0
xz=5.2.6=h5eee18b_0
yarl=1.8.1=pypi_0
zipp=3.10.0=pypi_0
zlib=1.2.12=h5eee18b_3
zstd=1.5.0=ha95c52a_0
